First commit, lots of stuff:
1. Changed all group names from 'VM' and 'VMUI' to more organized 'UserInterface', 'Skins', and 'Icons, and accordingly reorganized all sounds and textures to appropriate folders
2. Added DeusEx.Upkg
3. Added in 4 more normal ricochet sounds, plus material based ricochet sounds that are layered on top for wood, glass and dirt
4. Semi working material based debris spawning for darts/shurikens

Currently both of the debris spawning code are slightly busted, because although the trace checks for walls, for some reason it doesn't check for floors and only one can be done at a time. Very strange as to why, and I have an idea of what to do that might not be pretty.